Analyzer: Resolve the relations (tables and views) against the search_path.
We now support a list of schemas in the search_path (pg_catalog being the
first, implicit, one in the list). When analyzing statements we now resolve
the relation's schema (if not specified) against the configured search_path
by iterating over the list of schemas and yielding the first one that contains
the relation. The `create table` statement is the only exception to this rule
which, if the schema is not specified, will use the default schema (ie. the
first one in the search_path)

In case we don't find a matching schema we will report this using a schemaless
message in the form of "Relation x unknown".

This also drops the PostgresITTest.testCustomSchemaAndAnalyzerFailure test
as we'll not report that schema is unknown but relation unknown.